circbuf: Move from mm/circbuf to libs/libc/misc
so that it can be used by userspace program.
