+++ arch/xtensa/src/esp32s2/esp32s2_i2s.c
@@ -42,7 +42,7 @@
 #include <nuttx/clock.h>
 #include <nuttx/semaphore.h>
 #include <nuttx/mqueue.h>
-#include <nuttx/mm/circbuf.h>
+#include <nuttx/circbuf.h>
 #include <nuttx/audio/audio.h>
 #include <nuttx/audio/i2s.h>
