+++ arch/xtensa/src/esp32s3/esp32s3_i2s.c
@@ -45,7 +45,7 @@
 #include <nuttx/semaphore.h>
 #include <nuttx/spinlock.h>
 #include <nuttx/mqueue.h>
-#include <nuttx/mm/circbuf.h>
+#include <nuttx/circbuf.h>
 #include <nuttx/audio/audio.h>
 #include <nuttx/audio/i2s.h>
