+++ drivers/rc/lirc_dev.c
@@ -34,7 +34,7 @@
 
 #include <nuttx/kmalloc.h>
 #include <nuttx/mutex.h>
-#include <nuttx/mm/circbuf.h>
+#include <nuttx/circbuf.h>
 #include <nuttx/rc/lirc_dev.h>
 
 /****************************************************************************
