+++ include/nuttx/circbuf.h
@@ -1,5 +1,5 @@
 /****************************************************************************
- * mm/circbuf/circbuf.c
+ * include/nuttx/circbuf.h
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -20,6 +20,9 @@
  *
  ****************************************************************************/
 
+#ifndef __INCLUDE_NUTTX_CIRCBUF_H
+#define __INCLUDE_NUTTX_CIRCBUF_H
+
 /* Note about locking: There is no locking required while only one reader
  * and one writer is using the circular buffer.
  * For multiple writer and one reader there is only a need to lock the
@@ -31,20 +34,36 @@
  * Included Files
  ****************************************************************************/
 
-#include <nuttx/config.h>
-
-#include <assert.h>
-
-#include <nuttx/kmalloc.h>
-#include <nuttx/mm/circbuf.h>
+#include <stdbool.h>
+#include <sys/types.h>
 
 /****************************************************************************
- * Private Types
+ * Public Types
  ****************************************************************************/
 
+/* This structure describes circular buffer */
+
+struct circbuf_s
+{
+  FAR void *base;     /* The pointer to buffer space */
+  size_t    size;     /* The size of buffer space */
+  size_t    head;     /* The head of buffer space */
+  size_t    tail;     /* The tail of buffer space */
+  bool      external; /* The flag for external buffer */
+};
+
 /****************************************************************************
- * Public Functions
+ * Public Function Prototypes
  ****************************************************************************/
+
+#undef EXTERN
+#if defined(__cplusplus)
+#define EXTERN extern "C"
+extern "C"
+{
+#else
+#define EXTERN extern
+#endif
 
 /****************************************************************************
  * Name: circbuf_init
@@ -65,29 +84,8 @@
  *
  ****************************************************************************/
 
-int circbuf_init(FAR struct circbuf_s *circ, FAR void *base, size_t bytes)
-{
-  DEBUGASSERT(circ);
-  DEBUGASSERT(!base || bytes);
-
-  circ->external = !!base;
-
-  if (!base && bytes)
-    {
-      base = kmm_malloc(bytes);
-      if (!base)
-        {
-          return -ENOMEM;
-        }
-    }
-
-  circ->base = base;
-  circ->size = bytes;
-  circ->head = 0;
-  circ->tail = 0;
-
-  return 0;
-}
+int circbuf_init(FAR struct circbuf_s *circ,
+                  FAR void *base, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_resize
@@ -104,61 +102,7 @@ int circbuf_init(FAR struct circbuf_s *circ, FAR void *base, size_t bytes)
  *
  ****************************************************************************/
 
-int circbuf_resize(FAR struct circbuf_s *circ, size_t bytes)
-{
-  FAR void *tmp = NULL;
-  size_t len = 0;
-
-  DEBUGASSERT(circ);
-  DEBUGASSERT(!circ->external);
-  if (bytes == circ->size)
-    {
-      return 0;
-    }
-
-  if (bytes)
-    {
-      tmp = kmm_malloc(bytes);
-      if (!tmp)
-        {
-          return -ENOMEM;
-        }
-
-      len = circbuf_used(circ);
-      if (bytes < len)
-        {
-          circbuf_skip(circ, len - bytes);
-          len = bytes;
-        }
-
-      circbuf_read(circ, tmp, len);
-    }
-
-  kmm_free(circ->base);
-
-  circ->base = tmp;
-  circ->size = bytes;
-  circ->head = len;
-  circ->tail = 0;
-
-  return 0;
-}
-
-/****************************************************************************
- * Name: circbuf_reset
- *
- * Description:
- *   Remove the entire circular buffer content.
- *
- * Input Parameters:
- *   circ  - Address of the circular buffer to be used.
- ****************************************************************************/
-
-void circbuf_reset(FAR struct circbuf_s *circ)
-{
-  DEBUGASSERT(circ);
-  circ->head = circ->tail = 0;
-}
+int circbuf_resize(FAR struct circbuf_s *circ, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_uninit
@@ -170,64 +114,19 @@ void circbuf_reset(FAR struct circbuf_s *circ)
  *   circ  - Address of the circular buffer to be used.
  ****************************************************************************/
 
-void circbuf_uninit(FAR struct circbuf_s *circ)
-{
-  DEBUGASSERT(circ);
-
-  if (!circ->external)
-    {
-      kmm_free(circ->base);
-    }
-
-  memset(circ, 0, sizeof(*circ));
-}
+void circbuf_uninit(FAR struct circbuf_s *circ);
 
 /****************************************************************************
- * Name: circbuf_size
+ * Name: circbuf_reset
  *
  * Description:
- *   Return size of the circular buffer.
+ *   Remove the entire circular buffer content.
  *
  * Input Parameters:
  *   circ  - Address of the circular buffer to be used.
  ****************************************************************************/
 
-size_t circbuf_size(FAR struct circbuf_s *circ)
-{
-  DEBUGASSERT(circ);
-  return circ->size;
-}
-
-/****************************************************************************
- * Name: circbuf_used
- *
- * Description:
- *   Return the used bytes of the circular buffer.
- *
- * Input Parameters:
- *   circ  - Address of the circular buffer to be used.
- ****************************************************************************/
-
-size_t circbuf_used(FAR struct circbuf_s *circ)
-{
-  DEBUGASSERT(circ);
-  return circ->head - circ->tail;
-}
-
-/****************************************************************************
- * Name: circbuf_space
- *
- * Description:
- *   Return the remaining space of the circular buffer.
- *
- * Input Parameters:
- *   circ  - Address of the circular buffer to be used.
- ****************************************************************************/
-
-size_t circbuf_space(FAR struct circbuf_s *circ)
-{
-  return circbuf_size(circ) - circbuf_used(circ);
-}
+void circbuf_reset(FAR struct circbuf_s *circ);
 
 /****************************************************************************
  * Name: circbuf_is_init
@@ -239,25 +138,7 @@ size_t circbuf_space(FAR struct circbuf_s *circ)
  *   circ  - Address of the circular buffer to be used.
  ****************************************************************************/
 
-bool circbuf_is_init(FAR struct circbuf_s *circ)
-{
-  return !!circ->base;
-}
-
-/****************************************************************************
- * Name: circbuf_is_empty
- *
- * Description:
- *   Return true if the circular buffer is empty.
- *
- * Input Parameters:
- *   circ  - Address of the circular buffer to be used.
- ****************************************************************************/
-
-bool circbuf_is_empty(FAR struct circbuf_s *circ)
-{
-  return !circbuf_used(circ);
-}
+bool circbuf_is_init(FAR struct circbuf_s *circ);
 
 /****************************************************************************
  * Name: circbuf_is_full
@@ -269,10 +150,55 @@ bool circbuf_is_empty(FAR struct circbuf_s *circ)
  *   circ  - Address of the circular buffer to be used.
  ****************************************************************************/
 
-bool circbuf_is_full(FAR struct circbuf_s *circ)
-{
-  return !circbuf_space(circ);
-}
+bool circbuf_is_full(FAR struct circbuf_s *circ);
+
+/****************************************************************************
+ * Name: circbuf_is_empty
+ *
+ * Description:
+ *   Return true if the circular buffer is empty.
+ *
+ * Input Parameters:
+ *   circ  - Address of the circular buffer to be used.
+ ****************************************************************************/
+
+bool circbuf_is_empty(FAR struct circbuf_s *circ);
+
+/****************************************************************************
+ * Name: circbuf_size
+ *
+ * Description:
+ *   Return size of the circular buffer.
+ *
+ * Input Parameters:
+ *   circ  - Address of the circular buffer to be used.
+ ****************************************************************************/
+
+size_t circbuf_size(FAR struct circbuf_s *circ);
+
+/****************************************************************************
+ * Name: circbuf_used
+ *
+ * Description:
+ *   Return the used bytes of the circular buffer.
+ *
+ * Input Parameters:
+ *   circ  - Address of the circular buffer to be used.
+ ****************************************************************************/
+
+size_t circbuf_used(FAR struct circbuf_s *circ);
+
+/****************************************************************************
+ * Name: circbuf_space
+ *
+ * Description:
+ *   Return the remaining space of the circular buffer.
+ *
+ * Input Parameters:
+ *   circ  - Address of the circular buffer to be used.
+ ****************************************************************************/
+
+size_t circbuf_space(FAR struct circbuf_s *circ);
 
 /****************************************************************************
  * Name: circbuf_peekat
@@ -296,42 +222,7 @@ bool circbuf_is_full(FAR struct circbuf_s *circ)
  ****************************************************************************/
 
 ssize_t circbuf_peekat(FAR struct circbuf_s *circ, size_t pos,
-                       FAR void *dst, size_t bytes)
-{
-  size_t len;
-  size_t off;
-
-  DEBUGASSERT(circ);
-
-  if (!circ->size)
-    {
-      return 0;
-    }
-
-  if (circ->head - pos > circ->head - circ->tail)
-    {
-      pos = circ->tail;
-    }
-
-  len = circ->head - pos;
-  off = pos % circ->size;
-
-  if (bytes > len)
-    {
-      bytes = len;
-    }
-
-  len = circ->size - off;
-  if (bytes < len)
-    {
-      len = bytes;
-    }
-
-  memcpy(dst, (FAR char *)circ->base + off, len);
-  memcpy((FAR char *)dst + len, circ->base, bytes - len);
-
-  return bytes;
-}
+                       FAR void *dst, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_peek
@@ -354,10 +245,7 @@ ssize_t circbuf_peekat(FAR struct circbuf_s *circ, size_t pos,
  ****************************************************************************/
 
 ssize_t circbuf_peek(FAR struct circbuf_s *circ,
-                     FAR void *dst, size_t bytes)
-{
-  return circbuf_peekat(circ, circ->tail, dst, bytes);
-}
+                      FAR void *dst, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_read
@@ -380,16 +268,7 @@ ssize_t circbuf_peek(FAR struct circbuf_s *circ,
  ****************************************************************************/
 
 ssize_t circbuf_read(FAR struct circbuf_s *circ,
-                     FAR void *dst, size_t bytes)
-{
-  DEBUGASSERT(circ);
-  DEBUGASSERT(dst || !bytes);
-
-  bytes = circbuf_peek(circ, dst, bytes);
-  circ->tail += bytes;
-
-  return bytes;
-}
+                      FAR void *dst, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_skip
@@ -397,7 +276,7 @@ ssize_t circbuf_read(FAR struct circbuf_s *circ,
  * Description:
  *   Skip data from the circular buffer.
  *
- * Note :
+ * Note:
  *   That with only one concurrent reader and one concurrent writer,
  *   you don't need extra locking to use these api.
  *
@@ -410,23 +289,7 @@ ssize_t circbuf_read(FAR struct circbuf_s *circ,
  *   A negated errno value is returned on any failure.
  ****************************************************************************/
 
-ssize_t circbuf_skip(FAR struct circbuf_s *circ, size_t bytes)
-{
-  size_t len;
-
-  DEBUGASSERT(circ);
-
-  len = circbuf_used(circ);
-
-  if (bytes > len)
-    {
-      bytes = len;
-    }
-
-  circ->tail += bytes;
-
-  return bytes;
-}
+ssize_t circbuf_skip(FAR struct circbuf_s *circ, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_write
@@ -434,7 +297,7 @@ ssize_t circbuf_skip(FAR struct circbuf_s *circ, size_t bytes)
  * Description:
  *   Write data to the circular buffer.
  *
- * Note :
+ * Note:
  *   That with only one concurrent reader and one concurrent writer,
  *   you don't need extra locking to use these api.
  *
@@ -449,38 +312,7 @@ ssize_t circbuf_skip(FAR struct circbuf_s *circ, size_t bytes)
  ****************************************************************************/
 
 ssize_t circbuf_write(FAR struct circbuf_s *circ,
-                      FAR const void *src, size_t bytes)
-{
-  size_t space;
-  size_t off;
-
-  DEBUGASSERT(circ);
-  DEBUGASSERT(src || !bytes);
-
-  if (!circ->size)
-    {
-      return 0;
-    }
-
-  space = circbuf_space(circ);
-  off = circ->head % circ->size;
-  if (bytes > space)
-    {
-      bytes = space;
-    }
-
-  space = circ->size - off;
-  if (bytes < space)
-    {
-      space = bytes;
-    }
-
-  memcpy((FAR char *)circ->base + off, src, space);
-  memcpy(circ->base, (FAR char *)src + space, bytes - space);
-  circ->head += bytes;
-
-  return bytes;
-}
+                       FAR const void *src, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_overwrite
@@ -505,49 +337,7 @@ ssize_t circbuf_write(FAR struct circbuf_s *circ,
  ****************************************************************************/
 
 ssize_t circbuf_overwrite(FAR struct circbuf_s *circ,
-                          FAR const void *src, size_t bytes)
-{
-  size_t overwrite = 0;
-  size_t skip = 0;
-  size_t space;
-  size_t off;
-
-  DEBUGASSERT(circ);
-  DEBUGASSERT(src || !bytes);
-
-  if (!circ->size)
-    {
-      return 0;
-    }
-
-  if (bytes > circ->size)
-    {
-      skip = bytes - circ->size;
-      src = (FAR const void *)((FAR char *)src + skip);
-      bytes = circ->size;
-    }
-
-  space = circbuf_space(circ);
-  if (bytes > space)
-    {
-      overwrite = bytes - space + skip;
-    }
-
-  circ->head += skip;
-  off = circ->head % circ->size;
-  space = circ->size - off;
-  if (bytes < space)
-    {
-      space = bytes;
-    }
-
-  memcpy((FAR char *)circ->base + off, src, space);
-  memcpy(circ->base, (FAR char *)src + space, bytes - space);
-  circ->head += bytes;
-  circ->tail += overwrite;
-
-  return overwrite;
-}
+                           FAR const void *src, size_t bytes);
 
 /****************************************************************************
  * Name: circbuf_get_writeptr
@@ -557,33 +347,14 @@ ssize_t circbuf_overwrite(FAR struct circbuf_s *circ,
  *
  * Input Parameters:
  *   circ  - Address of the circular buffer to be used.
- *   size  - Returns the maximum size that can be written consecutively
+ *   size  - Returns the maximum size that can be written consecutively.
  *
  * Returned Value:
  *   The write pointer of the circbuf.
  *
  ****************************************************************************/
 
-FAR void *circbuf_get_writeptr(FAR struct circbuf_s *circ, FAR size_t *size)
-{
-  size_t off;
-  size_t pos;
-
-  DEBUGASSERT(circ);
-
-  off = circ->head % circ->size;
-  pos = circ->tail % circ->size;
-  if (off >= pos)
-    {
-      *size = circ->size - off;
-    }
-  else
-    {
-      *size = pos - off;
-    }
-
-  return (FAR char *)circ->base + off;
-}
+FAR void *circbuf_get_writeptr(FAR struct circbuf_s *circ, FAR size_t *size);
 
 /****************************************************************************
  * Name: circbuf_get_readptr
@@ -600,26 +371,7 @@ FAR void *circbuf_get_writeptr(FAR struct circbuf_s *circ, FAR size_t *size)
  *
  ****************************************************************************/
 
-FAR void *circbuf_get_readptr(FAR struct circbuf_s *circ, size_t *size)
-{
-  size_t off;
-  size_t pos;
-
-  DEBUGASSERT(circ);
-
-  off = circ->head % circ->size;
-  pos = circ->tail % circ->size;
-  if (pos > off)
-    {
-      *size = circ->size - pos;
-    }
-  else
-    {
-      *size = off - pos;
-    }
-
-  return (FAR char *)circ->base + pos;
-}
+FAR void *circbuf_get_readptr(FAR struct circbuf_s *circ, FAR size_t *size);
 
 /****************************************************************************
  * Name: circbuf_writecommit
@@ -635,11 +387,7 @@ FAR void *circbuf_get_readptr(FAR struct circbuf_s *circ, size_t *size)
  *
  ****************************************************************************/
 
-void circbuf_writecommit(FAR struct circbuf_s *circ, size_t writtensize)
-{
-  DEBUGASSERT(circ);
-  circ->head += writtensize;
-}
+void circbuf_writecommit(FAR struct circbuf_s *circ, size_t writtensize);
 
 /****************************************************************************
  * Name: circbuf_readcommit
@@ -655,8 +403,10 @@ void circbuf_writecommit(FAR struct circbuf_s *circ, size_t writtensize)
  *
  ****************************************************************************/
 
-void circbuf_readcommit(FAR struct circbuf_s *circ, size_t readsize)
-{
-  DEBUGASSERT(circ);
-  circ->tail += readsize;
+void circbuf_readcommit(FAR struct circbuf_s *circ, size_t readsize);
+
+#undef EXTERN
+#if defined(__cplusplus)
 }
+#endif
+#endif
